fix int -> intptr_t to be consistent with Ronan previous updates (new PTR_NIL default value and index types...).
I do not really understand the point of turning this index into a 64bit
integer, while its value is expected to be in the 0 .. 1970 range.


git-svn-id: http://svn.cri.ensmp.fr/svn/linear/trunk@1259 8716bff9-3cf5-0310-8878-b27a67f92df8

+++ src/sc/sc_simplexe_feasibility.c
@@ -489,7 +489,7 @@ typedef struct
     int numero; 
     int hash ;
     Value val ;
-    int succ ;
+    intptr_t succ ;
 } hashtable_t;
 
 /* For debugging: */
